requirements.txt
requests
pandas
python-dotenv
openpyxl
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
